requirements.txt
aiofiles==23.1.0
aiogram==2.25.1
aiohttp==3.8.4
aioschedule==0.5.2
aiosignal==1.3.1
anyio==3.7.0
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.9.1
beautifulsoup4==4.12.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
defusedxml==0.7.1
emoji==2.6.0
exceptiongroup==1.1.2
fake-useragent==1.1.3
filelock==3.12.2
frozenlist==1.3.3
h11==0.14.0
httpcore==0.17.2
httpx==0.24.1
idna==3.4
lxml==4.9.2
magic-filter==1.0.9
multidict==6.0.4
outcome==1.2.0
pyaes==1.6.1
pyasn1==0.5.0
pycparser==2.21
pydantic==2.0.1
PySocks==1.7.1
python-dotenv==1.0.0
python-telegram-bot==20.3
pytz==2023.3
requests==2.31.0
rfc3986==2.0.0
rsa==4.9
schedule==1.2.0
selenium==4.10.0
sniffio==1.3.0
snscrape==0.7.0.20230622
sortedcontainers==2.4.0
soupsieve==2.4.1
trio==0.22.1
trio-websocket==0.10.3
typing_extensions==4.7.1
untangle==1.2.1
urllib3==2.0.3
wsproto==1.2.0
xmltodict==0.13.0
yarl==1.9.2

Flask~=2.3.2